Refactor yggdrasil controllers
