+++ backend/apps/xlmine/routes/yggdrasil.py
@@ -1,10 +1,19 @@
 # xlmine/routes/yggdrasil.py
 from django.urls import path
 
-from apps.xlmine.controllers.yggdrasil import (
-    profile_view, has_joined_view, join_server_view,
-    signout_view, invalidate_view, validate_view,
-    refresh_view, authenticate_view, base, auth_server
+from apps.xlmine.controllers.auth import (
+    base,
+    auth_server,
+    authenticate_view,
+    refresh_view,
+    validate_view,
+    invalidate_view,
+    signout_view,
+)
+from apps.xlmine.controllers.session import (
+    join_server_view,
+    has_joined_view,
+    profile_view,
 )
 
 urlpatterns = [
